% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distanceTransform}
\alias{distanceTransform}
\alias{distanceTransform.default}
\title{Distance transforms}
\usage{
distanceTransform(x, ...)

\method{distanceTransform}{default}(x, ...)
}
\arguments{
\item{x}{Any object. For the default method, this must be coercible to an
array.}

\item{\dots}{Additional arguments to methods.}
}
\value{
An array of the same dimension as the original, whose elements give
  the Euclidean distance from that element to the nearest "on" element in
  the original.
}
\description{
The Euclidean distance transform produces an array like its argument, but
with element values representing the Euclidean distance to the nearest
nonzero element. The input is treated as logically binary, with all nonzero
values treated as "on", and all zeroes as "off".
}
\examples{
x <- c(0,0,1,0,0,0,1,1,1,0,0)
distanceTransform(x)
}
\author{
Jon Clayden <code@clayden.org>
}
