% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{symmetric}
\alias{symmetric}
\title{Check for a symmetric array}
\usage{
symmetric(x)
}
\arguments{
\item{x}{An object that can be coerced to a numeric array.}
}
\value{
A logical value indicating whether the array is symmetric or not.
}
\description{
This function checks whether a numeric array is symmetric, in the sense of
transposition. This is tested by comparing the reversed vectorised array to
the unreversed equivalent.
}
\author{
Jon Clayden <code@clayden.org>
}
